requirement.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.4
django-import-export==3.1.0
et-xmlfile==1.1.0
Faker==15.3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
python-dateutil==2.8.2
python-environ==0.4.54
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
xlrd==2.0.1
xlwt==1.3.0
